requirements.txt
azure-ai-documentintelligence
azure-identity
camelot-py
importlib-metadata
jaraco.collections
langchain-community
langchain-openai
mammoth
markitdown
olefile
pip-chill
platformdirs
pydub
python-pptx
scrapy
speechrecognition
xlrd
youtube-transcript-api
zope.interface
